Refactor Mongo DB interface ready for new tune.
This is identical in behaviour to the previous version, but refactored in order to make it easy to alter the unique key for a tune on the database.  We are about to add a key 'tid' which contains the existing _id content and revert to using Mongo's GUID-based _id.

Fixed error in pseudoMongoTransaction
Incremented Welcome version to 1.1.3 - also in build.sbt