Allow tunes to be replaced without prior deletion
The new tune URL now has upsert semantics, unless the new submitter is different from the original submitter in which case a 'tune exists' error is returned. https://github.com/newlandsvalley/musicrest/issues/5.

Ensure comments pertaining to a tune are removed if a tune is deleted.

Upgrade to Spray 1.1-RC4