Added tool to add a user remotely via HTTP
This is useful if we'd like to be able to use the functionality where the administrator can insert a pre-registered user without having to log in to the target machine.

Also, added some more delays to the tests to give Mongo time to keep up, and moved the exception handling and logging of routes so that they protect the concatenated routes and do not protect each route individually.  This means that the user route is protected for the first time and should give back meaningful error messages.