Use eps as intermediate format in image generation
The scripts producing pdf and png currently use ps as an intermediate representation.  This has the disadvantage that it is predominantly aimed at text formats meaning that large images might be split into seperate files for separate pages and so become virtually unretrievable by the software expecting a single page in all circumstances.  eps is aimed more at images and should retain the single page output.