Upgrade to scala 2.11.7, java 1.8.0
This has proved to be a major headache.  It's been two years since the last changes to this repository, and it has been relatively difficult to find the consistent set of dependencies (particularly with transitive dependencies to scalaz).  In this time, a good deal has changed:

The Casbah Mongo API is substantially different:  MongoConnection has become MongoClient; there are no pseudo-transaction primitives any longer; ensureIndex is deprecated in favour of createIndex; Casbah itself is now deprecated in facvour of a (presumably much better) new API.  Casbah is retained for the time being, though.

Scalaz has realized its mistake of calling Validation a monad.  As alomost the entire API is expressed in terms of Validations rather than disjunctions, this has been a distressing change to make. In the most part by retaining Validation and translating to a disjunction during for comprehensions and bacj when they've finished.

Specs2 has altered such that nearly every test failed.  It turns out that to test APIs returning Validations is no longer so easy, because a Validation failure that produces a Specs Failure is now a different type from a success expressed in terms of a match and can no longer be unified, so all these tests failed to compile.